migrate panel types; migrate more contextMenu types
