+++ src/background/contextMenus.test.ts
@@ -20,10 +20,7 @@ import {
   ensureContextMenu,
 } from "@/background/contextMenus";
 import extensionPointRegistry from "@/starterBricks/registry";
-import {
-  fromJS,
-  type MenuDefinition,
-} from "@/starterBricks/contextMenu/contextMenu";
+import { fromJS } from "@/starterBricks/contextMenu/contextMenu";
 import * as backgroundApi from "@/background/messenger/api";
 import { type StarterBrickConfig } from "@/starterBricks/types";
 import { type ModComponentBase } from "@/types/modComponentTypes";
@@ -32,7 +29,10 @@ import { TEST_setContext } from "webext-detect-page";
 import { modComponentFactory } from "@/testUtils/factories/modComponentFactories";
 import { starterBrickConfigFactory } from "@/testUtils/factories/modDefinitionFactories";
 import { getPlatform } from "@/platform/platformContext";
-import { type ContextMenuConfig } from "@/starterBricks/contextMenu/types";
+import {
+  type MenuDefinition,
+  type ContextMenuConfig,
+} from "@/starterBricks/contextMenu/types";
 
 TEST_setContext("background");
